feat: Switch to using account's transaction list
Using the account's transaction list (`GET ​/users​/{userId}​/accounts​/{accountId}​/transactions`) instead of the user's transaction list (`GET ​/users​/{userId}​/transactions`) is more in line with how we want to have 3rd party developers use the API. There are performance benefits to using the account's transaction list rather than the user's transaction list.

In addition to using a different endpoint, this commit also reformats the output  of the `/accountsAndTransactions` route so it is visually clear that particular transactions belong to a particular account. The `async function getTransactions()` was modified to accept an `accountId` in order to retrieve the transactions for that `accountId`.