feat: add scopes that are necessary to get data (#27)
* fix: crash when no Transactions found

It's entirely possible for an Account to have zero Transactions. My usage was breaking due to setting up a Bill Pay account (which had no transactions).

* feat: add scopes that are necessary to get data

Now that there is enforcement of the OAuth scopes for API endpoints, we needed to update this example to use the scopes for retrieving Accounts and Transactions data.