fix: use `keepSessionInfo` to maintain session
The passport.js changes in `0.6.0` have breaking changes related to protecting against "Session Fixation".
- https://github.com/jaredhanson/passport/pull/900
- https://medium.com/passportjs/fixing-session-fixation-b2b68619c51d

The assumption for the fix in this commit is that our example project here only has the session storage as its storage mechanism, so we're not quite vulnerable to the same thing since the storage goes away when the local project is stopped.