Reduce friction points (#45)
* docs: redirection of '/' to '/hello'

Folks sometimes click the log statement "Server listening on http://localhost:8080", which navigates them to the '/' route, which then gets redirected to the '/hello' route.

This was already handled, functionally, but didn't have comments explaining what was happening. So we add some in this commit.

* feat: Unique Customer Identifier (commented out)

This provides a line of code--commented out--and docs to explain how to retrieve the Unique Customer Identifier which is a Restricted Claim.

Since this is a Restricted Claim, we don't want folks to get confused by the fact that the claim isn't returned when the claim isn't turned on for the External Application.

The hope here is that the developer can reach out to the administrator at the Financial Institution to get the restricted claim turned on for their External Application. Then, they can uncomment this line and see the Unique Customer Identifier in the logs.